+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Show popups in Wellness ToDo and Rings just once [#2071](https://github.com/rokwire/illinois-app/issues/2071).
 - Fix blue color in Wellness ToDo categories [#2040](https://github.com/rokwire/illinois-app/issues/2040).
+### Changed
+- Create default order of Favorites panel widgets from extended FlexUI content [#2076](https://github.com/rokwire/illinois-app/issues/2076).
 
 ## [4.1.0] - 2022-07-20
 ### Changed